[ENH] Refactor plotting separate image (#5481)
