+++ nilearn/plotting/image/tests/test_plot_anat.py
