+++ nilearn/plotting/image/tests/test_plot_carpet.py
