+++ nilearn/plotting/image/tests/test_plot_connectome.py
