+++ nilearn/plotting/image/tests/test_plot_glass_brain.py
