+++ nilearn/plotting/image/tests/test_plot_img.py
