+++ nilearn/plotting/image/tests/test_plot_markers.py
