+++ nilearn/plotting/image/tests/test_plot_prob_atlas.py
