+++ nilearn/plotting/image/tests/test_plot_roi.py
