+++ nilearn/plotting/image/tests/test_plot_stat_map.py
