+++ nilearn/plotting/tests/test_utils.py
@@ -1,5 +1,5 @@
 """Tests for private function \
-   nilearn.plotting.img_plotting.get_colorbar_and_data_ranges.
+   nilearn.plotting.image.img_plotting.get_colorbar_and_data_ranges.
 """
 
 import numpy as np
